# projects/picker/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Picker Server API",
    "description": "Picker 服务器 API 文档",
    "contact": {
      "name": "API Support",
      "email": "openpicklabs@hotmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发服务器"
    },
    {
      "url": "https://api.openpick.org",
      "description": "生产服务器"
    }
  ],
  "paths": {
...
    "/download": {
      "get": {
        "tags": [
          "download"
        ],
        "summary": "下载Picker文件",
        "description": "使用下载令牌下载已购买的Picker文件",
        "operationId": "download",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "下载令牌",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "文件下载成功",
            "content": {
              "application/octet-stream": {

              }
            }
          },
          "401": {
            "description": "令牌无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "订单、Picker或文件不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "Invalid input parameters"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "请输入 Bearer Token，格式：Bearer \u003Cyour-token\u003E"
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}